Fix negative Best Segment Times
The comparison time fixing algorithm modifies the actual value, but uses a copy created before in the remainder of the algorithm. This results in an outdated value being used, causing negative values to be assigned to the best segment time. This fixing algorithm should probably be a method on the segment instead, so that it can access the internal state in a cleaner way.