Better Caching for Splits Component
Instead of returning a changed icon whenever scrolling happens, each
segment now has an associated segment index that'll help you track
movement of segments. The icon changes are now also not associated to
the segments themselves anymore, but rather are a list of changes as
pairings of segment ids and their changed icon. The component now
assumes that any icon is always known until either the component
remounts or it actually changes. So if an icon scrolls out of view and
back again, it is assumed to be already cached by the UI logic and
therefore won't be part of the icon changes again.