Document and Test the Panicking Behavior of Run
There's a lot of methods meant for the Run Editor and the Timer, both of which assume that the Run always has one segment. This is however not the case when a Run is used directly, so this commit properly documents when you are not supposed to use certain methods with a Run.

I'm not entirely happy with certain methods currently panicking, when their name really doesn't imply it at all, like the fix_splits method. You'd think you can always call it.