Fix reset and set attempt as PB logic
https://github.com/LiveSplit/LiveSplit/issues/1077 shows that the
Attempt Ended time can sometimes be the default value instead of the
proper time. This is caused by saving a run as a PB before the run has
finished in the original LiveSplit. This didn't really happen in
livesplit-core, but the new reset_run_and_set_attempt_as_pb logic has to
be ported to livesplit-core regardless to keep them in sync.