Fix the negative Total Playtime Bug
The original LiveSplit had a bug that allowed for the ended timestamp in
the attempt history to be before the started timestamp. This messes up
the Total Playtime in both the original LiveSplit and livesplit-core.
This commit bumps the LSS file version to 1.7.1 and introduces a check
for all file versions below that in order to clean up those broken
attempt history elements in older splits files that potentially have
this bug. https://github.com/LiveSplit/LiveSplit/issues/1077 is now
resolved at least on livesplit-core's side.