Shrink images that are too large (#145)
This is necessary eventually. However it also comes with a few downsides such as longer compile times, larger binary size and longer parsing times of splits due to the shrinking. This introduces the `image-shrinking` feature that allows you to choose whether you want to compile this in. It is activated by default. The plan is to introduce this to the original LiveSplit and Splits I/O before this lands in LiveSplit One, so that we can reduce the amount of actively used splits in the wild that still need to be shrunk once LiveSplit One gets the feature. 

Closes #47