Improve Image Caching
For the native renderer it turns out that we need some kind of backwards
remounting. Essentially when the whole layout gets replaced, then the
new layout should realize that it got freshly mounted and should
potentially communicate that images that used to be shown by the old
layout shouldn't be shown anymore. By introducing an actual
CachedImageId type, we have a much cleaner way of managing this state
machine. It now has an uncached state, which is both the initial state
and the state after remounting. This way freshly mounted layouts always
respond with an actual image change.