Initial implementation of the rendering
This commit adds a rendering module behind a new feature `rendering` to
livesplit-core. The renderer here takes a LayoutState and an
implementation of the Backend trait. The renderer takes the layout state
and instructs the backend to creates meshes, textures and to render them
out. By abstracting out the backend we can use this rendering on a lot
of platforms without needing to focus on a specific rendering framework.