Introduce a notion of predecessors for Sum of Best
Annotating each individual segment with the predecessor, that its sum
of best segments value is derived from, allows for traversing the actual
shortest path that leads to the total sum of best. Due to combined
segments there is the potential that some segments end up as leaves that
don't contribute to that shortest path. These are incredibly confusing
for the Best Segments comparison, as the segment time between two
segments in that comparison may not mean anything and could even be
negative. This caused a variety of issues, which is why we only show the
actual shortest path in the Best Segments comparison now.