Fix up the Continuous Integration
In order to run the rendering tests on all the platforms a few changes
needed to be made:

1. The floating point accuracy of i586 platforms is inadequate for
rendering the layouts with enough precision to end up with the same
checksums as on the other platforms. The tests are therefore disabled on
platforms that don't support MMX. However Rust can't currently check for
MMX, so we check for SSE in the meantime.

2. Big Endian Platforms can't properly decode zlib streams correctly
such as present in PNG images. The platforms are in the allowed to fail
list for now until the PR I submitted is merged and published.

3. Turns out we can actually support more targets on Windows, such as
the i568 target. Support for aarch64 is almost there as well, so we are
tracking that now.