Add Basic Subsplits Support in Splits Component
All the splits without any segment groups behave like they did before
and the segment groups now properly open up and close as they are being
entered. This is done by transforming the segment groups iterator and
implementing an iterator on top that flattens it with a specific segment
in focus that determines the group that is supposed to be open. After
this flattening almost all the original logic remains, but now with the
current segment index and the run's length instead being based on this
flattened segment group representation.