Save Segment Groups into the splits files
It seems like we should have them as a separate block in there, such
that old parsers will have no problem simply ignoring that new block and
parsing the splits regardless. This however has the disadvantage that
the groups can now overlap, which we will have to account for when
parsing.