Improve the Rendering
This improves the ellipsis algorithm for text that doesn't fit. Before
we used to start cutting off text once the current width + the width of
the ellipsis becomes too large. However the actual remaining glyphs may
have fit there perfectly well. This is now fixed with the new algorithm.
It allocates a whole lot of space on the stack for buffering the glyphs,
but somehow that didn't seem to affect the benchmarks negatively. In
fact the benchmarks looked a tiny bit faster for some reason. Also with
the new ellipsis code, abbreviations for all the different info text
component keys are now supported too. Additionally the title component
now centers its single line text.

Fixes #171
Fixes #172