Implement Horizontal Layouts
This commit implements horizontal layouts for livesplit-core. The
horizontal layouts in the original LiveSplit were kind of weird in how
when changing the height of the timer, instead of the font sizes
adjusting, like in vertical mode, there would be a weird, really ugly
gap. Therefore there was essential only one predefined aspect ratio that
would look good. This time though, instead we use kind of a hybrid
approach for horizontal mode. There is now a fixed height for all the
horizontal layouts in component space. It's the two row height that we
also use in vertical mode. By using a fixed height, there never is any
gap. We then distribute the total width to the individual components
based on a preferred width they choose. They don't actually get their
preferred width, it only serves as a factor of how large the components
should be compared to each other. So similar to vertical mode,
horizontal mode now also has a varying width per component. This is
another feature that we did not have before. Is something cut off? Just
make the window wider and you'll give more space to the components,
allowing them to show text and other content, without it being cut off
as much.

Closes #173