Use proper spacing between rows when rendering
We positioned the second row pretty arbitrarily before. Now it is
positioned exactly one line below the one above. This however has
massive impact on everything and even redefines the two row height,
which is the total height we use in horizontal mode. So this has a huge
impact. However this should make everything more consistent and it is
clearly defined now, so it shouldn't change again in the future.