Update various dependencies
Looks like CI has been failing and there's been lots of updates for
various dependencies. So this commit fixes the tests and updates to the
latest versions of all the dependencies. Also to make it easier to see
what changed in our rendering tests, there's an automatic diff image
written out now for every test that is failing.