Port the PB Chance Component
This ports the PB Chance component from the original LiveSplit. Unlike
the original component, this uses a different algorithm though. The
livesplit-core version uses the same algorithm as the goal comparison
generator and the Balanced PB comparison. The percentile chosen by the
Balanced PB to reach the PB is the PB chance. Resets are ignored by this
algorithm entirely.

Resolves #208