Fix reordering bug when removing Run Metadata Variables
So far we have always used IndexMap's .remove() which is now considered
deprecated because it's not clear whether it shifts the elements or
swaps in one. IndexMap is meant to have stable indices, so it actually
does a swapping remove. This is not actually what we wanted as it is
destructive to the order of the variables, which is more important in
this case. This commit switches to the newly added .shift_remove() and
adds a test for it.