Switch to GitHub Actions
This commit switches our CI from Travis CI + AppVeyor to GitHub Actions.
There's a few minor problems still such as the Windows runners
sometimes not being able to checkout the commit or upload the artifacts.
This seems to be mitigated by running the Windows jobs first though.
Other problems include that the currently released version of cross
isn't compatible with GitHub Actions so we use a custom version in the
meantime.