Rework the WebAssembly Targets
With this Pull Request livesplit-core can now properly target the
WebAssembly System Interface (WASI) and wasm-bindgen. wasm-bindgen will
become more relevant for us in the future as we plan on using futures to
communicate with REST APIs such as speedrun.com and splits.io.

The binding generator still only works for the wasm32-unknown use case,
but that will have to change in the future.

Fixes #258