Unify Comparison Segment Time logic
Now everything is using the same logic for the comparison segment time.
However there is the notion of a combined and single segment time. The
detailed timer used single segment time, while the others used combined
segment time. It's not clear yet what we want. Also for the segment time
component, this should possibly even be a setting.