Separate the notion of combined and single segment times
We explicitly provide different functions for these to highlight the
differences. For the splits component we keep using the combined segment
times while for the Detailed Timer and Segment Time component we use
single segment times. The reason for this is that the Splits component
shows which splits have been skipped and are part of the combined
segment time, while the Detailed Timer and Segment Time component lack
this information.