Implement initial support for no_std
This implements initial support for no_std. Modules such as parsing,
saving and rendering are disabled and timing has to be manually provided
by registering a global handler. The rendering module is on its way to
no_std support as well, but not all Pull Requests to the necessary
dependencies are done yet.

See #223 for more information.