Remove emscripten
It's not feasible to maintain emscripten support anymore. We don't use
emscripten and it breaks all the time. Fixing it is hard because without
it being used anywhere, you don't know if the fix actually worked.
Additionally Edge is now using Chromium and we are long past supporting
Internet Explorer. If anyone ever intends to use a JS version of
livesplit-core again, they can use `wasm2js` to still compile the wasm
that we produce to JS. The output is quite large, but it should work.