Fix the Single Line Title Calculation
It assumed that for a single line that the game and category always
exist, but that is not always the case. We already handled that case,
just not for the single line title scenario. So this basically moves the
whole single line title calculation into the branch where we made sure
that both exist in the first place.