Properly read strings as UTF-8 in C# bindings
The encoding previously was locale dependent, so for some languages such
as Japanese the default encoding is not UTF-8. We now explicitly do the
UTF-8 decoding, so this should work across all setups now.

Additionally the string length is known by Rust, so we don't need to do
a loop looking for the nul-terminator anymore.