Expose the Skill Curve in the API
We've used the Balanced PB algorithm for various things by now: Balanced
PB itself, the goal generator, the PB chance and it theoretically could
also be used for the median segments and for calculating the simple sum
of best and worst segments. The underlying curve that is being sampled
here contains lots of interesting information, so it makes sense to
expose it for analysis purposes in the API. I dubbed it the "skill
curve" here, but the name is not necessarily final.