Implement updating the Layout States
So far we've always discarded the layout states, but since we want to
calculate layout states every frame, it makes a lot more sense to keep
the layout state around and just update it. All of the original API is
retained, so you are not forced to actually update your layout states.
However the code for creating a new layout state now initializes a
default state and updates it, instead of directly creating it correctly.
This might slightly slow down initial creation, but we don't want to
basically duplicate the same code across initial creation and updating.

By doing this an average frame now has 0 allocations, reallocations and
deallocations. We are seeing speedups of around 3x - 5x in the newly
added benchmarks.

This also introduces a new `ClearVec` type that behaves like a normal
`Vec` but when clearing it, instead of dropping all the elements, it
just keeps them around and clears them instead. So if you had a
`Vec<String>` and cleared it, all the Strings would get deallocated,
while a `ClearVec<String>` would keep them around and just set their
length back to 0. We use this in a few places to reduce the chances of
us needing to allocate new memory. It might make sense to move it into
its own crate eventually as this should be a useful type for any kind of
Rust code.