Introduce a Timer Snapshot Mechanism
This introduces the ability to create snapshots of the timer. At the
moment of snapshot creation the current time of the timer is evaluated.
The snapshot then represents the timer with the current time frozen at
that specific time. This allows layout state evaluation and other
calculations to work with a consistent timer that is entirely frozen,
allowing for more consistent results. However almost more importantly,
this reduces the amount of times we need to take time stamps from the
underlying system, improving the performance a bit.

![https://i.imgur.com/vDkucUu.png](https://i.imgur.com/vDkucUu.png)

In the web the performance improvement is even more significant. There
the layout state calculation improved from 8.5us to 5us.