Delay registering the Gamepad Hook
In the web we support gamepads. Unfortunately there's no callbacks or so
that we could use, so instead we have to use an interval where we poll
for button presses. This isn't super expensive, but still unnecessary if
you don't use a gamepad. Additionally Chrome has the problem that it
gets exclusive access to the gamepads and doesn't really yield control
back until the whole browser is closed. So you really don't want that
unless you really intend to use the gamepad for splitting. This is why
the gamepad interval is now only being set when the first gamepad button
is registered.