Allow accessing the mesh buffers as byte slices
When building a backend for the renderer you likely will end up needing
to upload the vertex and index buffers of the meshes to the GPU. Most
APIs want to consume byte slices for that. So this adds some convenience
methods that allow accessing the vertices and the indices as byte
slices.