Add Dummy Rendering Benchmark
Text shaping with harfbuzz is not really fast, and rustybuzz is
apparently still a bit slower than harfbuzz. So compared to rusttype
switching to rustybuzz may slow us down quite a bit. It would be good to
know how much time the renderer, without measuring any actual rendering,
is actually taking up on the CPU.