Switch Text Shaping to rustybuzz
rustybuzz is a more or less complete port of harfbuzz (state of the art)
text shaping engine to Rust. This means our font handling is now about
as good as the one found in a browser, except all the manual things you
may need to do on top such as using different fonts for different parts
of the text, because some glyphs are not available in the chosen font
and handling image based glyphs such as those that are being used for
emojis in some fonts. Overall though switching to rustybuzz seems to
already changed the appearance quite a bit. There is now proper kerning
(though I thought rusttype already did that, but apparently not), proper
ligature support and because this is state of the art, it also supports
tabular numbers for example which is now activated for all the numbers
that we are rendering. Unfortunately since rustybuzz does so much more
work, it also heavily slowed down text shaping, which we do on every
frame at the moment. So we really need to refactor the renderer to start
caching a lot more stuff.