Turn the renderer into a purely path based renderer
Our renderer always used paths internally but those quickly got turned
into meshes. It turns out that a lot of backends can actually directly
work with those paths, so splitting out the path rendering makes a lot
of sense. This also enables the use of `tiny-skia` for the software
renderer. However here is where it gets really surprising: It turns out
that `tiny-skia` is so fast, that it outperforms the GPU. And this is
without any optimizations. This isn't entirely surprising as Chrome
works in a very similar way where the GPU is mostly just used to compose
the different Skia layers together. So for now this also removes the
mesh based renderer entirely. We are going to investigate rendering into
different layers, so that's where a GPU backend may come back into play,
but not in the form of a mesh based renderer.

Additionally this adds support for color fonts using the `CPAL` and
`COLR` tables.