Don't read into uninitialized buffers
For performance reasons we used to read into uninitialized buffers.
However this is only safe if you can prove that you won't ever read from
the uninitialized memory, which isn't possible in our case since we
accept any reader that implements `std::io::Read`. There's no guarantee
that the reader only fills our buffer and doesn't also try to read from
it. It's unlikely that any reader will actually do that, but there's a
slight possibility that we can't prove away. They are currently trying
to address this in std, so that you can safely express this. For now we
just fall back to zero initializing the buffer. The Llanfair parsing
should be rare enough that this shouldn't matter all too much anyway.
There are FIXMEs in the code now that someone can eventually fix once
the safe reading into unitialized buffers is possible.

Resolves #400