Remove `palette` and `odds` as dependencies
We only needed `odds` for a single function that seems to be in `std` by
now and `palette` didn't provide all too much benefit for us either. We
initially used it for cleanly handling linear colors and gamma
compressed colors. However most UI frameworks don't actually implement
this correctly, so we ended up handling it like most UI frameworks do as
well, completely disregarding all correctness. It also is a slightly
heavy dependency because it has a proc macro dependency. So for these
reasons we will simply handle the colors ourselves.