Implement Hotkeys for macOS
This implements the global hotkeys for macOS. The code is mostly
based on:

https://github.com/kwhat/libuiohook/blob/f4bb19be8aee7d7ee5ead89b5a89dbf440e2a71a/src/darwin/input_hook.c#L1086

There's some stuff I left out, but it doesn't seem necessary.

Additionally I renamed all of the `NumPad*` variants of the
individual `KeyCode` enums to `Numpad*`. We probably really only
want a single enum that is used by all the platforms, so there's
no such disparities.