Implement Linked Layouts
This implements linked layouts that allow you to associate splits with
layouts. This means that a timer application can automatically open the
layout that's meant to be used for a particular splits file, improving
the user experience a lot. This already landed in the original
LiveSplit, so while the way it is stored in the XML is not ideal, in
particular with the way the default layout is annotated, we will have to
support it for backwards compatibility anyway. Further changes can be
made in the future.

Co-authored-by: wooferzfg <spapushin@gmail.com>