Use custom Instant implementation on Linux, macOS and iOS
We can't use std's Instant as it's insufficiently specified. It neither
guarantees "real time" nor does it guarantee measuring "uptime" (the
time the OS has been awake rather than suspended), meaning that you
can't actually rely on it in practice. In livesplit-core we definitely
want real time rather than uptime. The problem is mostly Linux. POSIX
intends for `CLOCK_MONOTONIC` to be real time, but this wasn't correctly
implemented in Linux and due to backwards compatibility concerns they
were never able to fix it properly. Thus `CLOCK_MONOTONIC` means uptime
on Linux whereas on other Unixes it means real time. They however
introduced `CLOCK_BOOTTIME` in the Linux kernel 2.6.39 which measures
real time. So the solution is to use this on all operating systems that
are based on the Linux kernel and fall back to `CLOCK_MONOTONIC` if the
kernel is too old and the syscall fails.

macOS and iOS actually do the right thing for `CLOCK_MONOTONIC` but Rust
actually doesn't use it on iOS and macOS, so we also need to use our
custom implementation for those too, but skip `CLOCK_BOOTTIME` as that
is Linux specific.

On Windows Instant currently measures real time, but we may need to use
a custom implementation for it as well in case this ever changes.