Add more docs for our `Instant` implementation
This removes emscripten from the list of Linux like operating systems as
it doesn't actually use the Linux kernel and instead brings its own JS
implementation that errors out if you call it with `CLOCK_BOOTTIME`.

Additionally we remove the atomic that caches whether `CLOCK_BOOTTIME`
works. There's a tiny range of Linux kernels that don't have it and are
still supported by Rust. While removing the atomic slows down these old
kernels, the slight overhead for all the recent Linux kernels just isn't
really worth it.

Also this adds a whole lot more documentation for what the situation
across all the different operating systems is, as well as the web.