Remove the `Short` Time Formatter
It turns out we don't actually really need it anywhere. It's the same as
the `SegmentTime` time formatter, except on `None` it pretends there to
be a value, which doesn't really make all that much sense. The run
editor even wraps it in an `EmptyWrapper` that replaces its `None` value
with an empty string, which is the same as just wrapping the
`SegmentTime` one.