Use `SegmentTime` Formatter for Segment Time Columns
When we introduced `Possible Time Save` columns we also added its
formatter. This formatter is the `SegmentTime` formatter. However the
actual column types that show various kinds of segment times still kept
using the `General` formatter that we use for split times. This changes
all of them to use the `SegmentTime` formatter.