Add an infallible way to resolve a `KeyCode`
In order to visualize a `KeyCode` in a user interface you need to
resolve it to a Keyboard layout specific representation. While I have a
work in progress implementation for the various platforms, there are a
couple of issues such as the web version needing to be async and the
Linux version requiring an X11 connection. Regardless, we always need to
have a fallback implementation that works if the OS somehow can't
resolve a specific KeyCode. This fallback implementation is simply based
on the standard US keyboard layout and can easily already be merged into
livesplit-core.