Switch to the `time` crate
The time crate was unusable when we first started writing
`livesplit-core`. The `chrono` crate was the go to crate for date and
time handling building on top of the very low level `time` crate. The
crate however got a new maintainer since then and became a really good
alternative to `chrono` which is still using the very old `0.1` version
of the `time` crate. Since then people also found in a CVE in the
`chrono` and old `time` crate versions. So this not only slightly
shrinks our dependency tree, but also resolves the CVE. This also
includes some slight performance improvements.