Introduce a per component cache in the renderer
At the moment the scene manager completely recalculates everything on
every update. This is really expensive, especially for the labels where
text shaping is apparently a pretty slow process. We want to do a lot
more caching to improve the performance. Here we introduce a per
component cache where each component can cache whatever it needs to. We
used to globally store icons in the scene manager to cache them, but now
these move into the individual component caches. Text shaping is still
done on each frame, but it's now possible to start moving these into the
component caches in further pull requests.