Introduce Virtual Key Fallback on Windows
Some key presses don't emit a scan code. I believe this is because of
certain drivers where the Keyboard itself doesn't emit the scan code and
instead the driver emits the virtual key to Windows. Windows provides an
API to convert between the virtual key and the scan code, so in case the
scan code is missing we attempt to convert the virtual key to the scan
code.

This also fixes the right shift and alt key. The alt key doesn't use
`WM_KEYDOWN` and instead uses `WM_SYSKEYDOWN`. And the right shift key
somehow is extended, which should not be the case.