Remove `euclid` and `ordered-float`
Both of these dependencies were barely needed. We only used the 2D
transformation matrix from `euclid`. However we don't actually need
rotations and also expose this type in the public API, so it's not ideal
to directly export a third party type like that as any major version
bump in `euclid` would also force us to do a major version bump as well.
This introduces a much more lightweight and also faster alternative that
we can expose in the API without any problems.

Additionally this removes `ordered-float`. In a lot of cases we can just
use the `TimeSpan` type instead of converting them to floats. This
actually has been the longest open issue in livesplit-core anyway, so
fixing the algorithms to use `TimeSpan` also removes the need for
sorting any floats... at least almost. In the few cases where we still
want to sort floats we can guarantee that no `NaN` is involved anyway,
so we can just do a simple comparison there without having to worry
about that. This should also be slightly faster anyway.