Use GDI to resolve fonts of the original LiveSplit
The original LiveSplit uses Windows Forms which internally uses the
fairly old GDI API. GDI's font identifiers encode styling information,
don't use the proper family name and are limited to 32 characters. This
makes it very hard for us to properly parse that information when
importing fonts from the original LiveSplit's layouts. However on
Windows we can directly talk to GDI to let it resolve the font. We can
then query more accurate information from the font by querying for the
`name` table, which contains the proper family name. This of course is
only available on Windows and only if the font can actually be found.
Otherwise we still provide the fallback algorithm to parse the font
information.