Introduce a Watchdog in the Auto Splitting Runtime
While the auto splitting crate itself provided a way to interrupt the
execution, it wasn't actually used by `livesplit-core`. This bumps the
version of `wasmtime` which actually removes the support for interrupt
handles, but provides an alternative system based on epochs. We now
provide our own interrupt handle that uses the epochs internally.
Additionally this handle is now actually used. There is now an
additional watchdog thread that expects a signal to be sent to it after
the time it takes for the next call to `update` to start (which is based
on the tick rate) and an additional 5 seconds it is allowed to execute.
If it doesn't stop executing after that time, the interrupt handle is
triggered and the module gets unloaded. Additionally the handle is used
in the `Drop` implementation of the runtime to ensure that it actually
shuts down.